requirements.txt
alembic==1.11.2
bcrypt==4.0.1
blinker==1.6.2
click==8.1.6
colorama==0.4.6
cssbeautifier==1.14.9
djlint==1.32.1
dnspython==2.4.1
EditorConfig==0.12.3
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-DotEnv==0.1.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsbeautifier==1.14.9
json5==0.9.14
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
six==1.16.0
SQLAlchemy==2.0.19
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.7.1
Werkzeug==2.3.6
WTForms==3.0.1
